+++ app/src/main/java/com/codesoom/assignment/errors/SessionControllerAdvice.java
@@ -1,0 +1,25 @@
+package com.codesoom.assignment.errors;
+
+import com.codesoom.assignment.dto.ErrorResponse;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+/**
+ * 세션과 연관해서 생기는 예외를 처리합니다.
+ */
+@RestControllerAdvice
+public class SessionControllerAdvice {
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(LoginFailException.class)
+    public ErrorResponse handleLoginFailException() {
+        return new ErrorResponse("Log-in failed");
+    }
+
+    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+    @ExceptionHandler(InvalidTokenException.class)
+    public ErrorResponse handleInvalidAccessTokenException() {
+        return new ErrorResponse("Invalid access token");
+    }
+}
